rename "store" to "context" folder
