+++ src/components/Navigation/Header/DropdownMenu/DropdownMenu.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, useContext } from "react";
 import { useTranslation } from "react-i18next";
 import { ReactComponent as LogoutIcon } from "../../../../assets/logout.svg";
-import { AppContext, Unit } from "../../../../store/app-context";
+import { AppContext, Unit } from "../../../../context/app-context";
 import Toggle from "../../../Shared/Toggle/Toggle";
 
 const DropdownMenu = forwardRef<HTMLDivElement>((_, ref) => {
