+++ src/components/Navigation/Header/Header.tsx
@@ -4,8 +4,8 @@ import { ReactComponent as MenuIcon } from "../../../assets/menu.svg";
 import { ReactComponent as RaspiBlitzMobileLogo } from "../../../assets/RaspiBlitz_Logo_Icon.svg";
 import { ReactComponent as RaspiBlitzLogo } from "../../../assets/RaspiBlitz_Logo_Main.svg";
 import { ReactComponent as RaspiBlitzLogoDark } from "../../../assets/RaspiBlitz_Logo_Main_Negative.svg";
-import { AppContext } from "../../../store/app-context";
-import { SSEContext } from "../../../store/sse-context";
+import { AppContext } from "../../../context/app-context";
+import { SSEContext } from "../../../context/sse-context";
 import DropdownMenu from "./DropdownMenu/DropdownMenu";
 
 const Header: FC = () => {
