+++ src/components/Shared/ListChannelModal/ListChannelModal.tsx
@@ -5,7 +5,7 @@ import { toast } from "react-toastify";
 import Message from "../../../container/Message/Message";
 import ModalDialog from "../../../container/ModalDialog/ModalDialog";
 import { LightningChannel } from "../../../models/lightning-channel";
-import { AppContext } from "../../../store/app-context";
+import { AppContext } from "../../../context/app-context";
 import { checkError } from "../../../util/checkError";
 import { instance } from "../../../util/interceptor";
 import { MODAL_ROOT } from "../../../util/util";
