+++ src/container/LoadingScreen/LoadingScreen.tsx
@@ -1,6 +1,6 @@
 import { FC, useContext } from "react";
 import LoadingSpinner from "../../components/Shared/LoadingSpinner/LoadingSpinner";
-import { AppContext } from "../../store/app-context";
+import { AppContext } from "../../context/app-context";
 import { ReactComponent as RaspiBlitzLogo } from "../../assets/RaspiBlitz_Logo_Main.svg";
 import { ReactComponent as RaspiBlitzLogoDark } from "../../assets/RaspiBlitz_Logo_Main_Negative.svg";
