+++ src/container/SetupContainer/SetupContainer.tsx
@@ -1,7 +1,7 @@
 import { FC, useContext } from "react";
 import { ReactComponent as MoonLogo } from "../../assets/moon.svg";
 import I18nDropdown from "../../components/Shared/I18nDropdown/I18nDropdown";
-import { AppContext } from "../../store/app-context";
+import { AppContext } from "../../context/app-context";
 
 type Props = {
   children?: React.ReactNode;
