+++ src/pages/Home/BitcoinCard.tsx
@@ -1,6 +1,6 @@
 import { FC, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { SSEContext } from "../../store/sse-context";
+import { SSEContext } from "../../context/sse-context";
 import { checkPropsUndefined } from "../../util/util";
 import LoadingBox from "../../components/Shared/LoadingBox/LoadingBox";
